vendor: Extend the AudioPolicy configuration for CodeAurora Audio
If we build the CodeAurora Foundation "downstream" based Audio HAL
instead of the one that is provided by Google in AOSP, we also want
to configure it to use its extensions.
Here we add the configuration extensions and routes variations
to use the advanced features provided by this kind of HAL.

To play nice and retain compatibility with both "CAF" and AOSP
provided HALs, add a TARGET_USES_AOSP_AUDIO_HAL barrier that
will exclude the extended configuration.