vendor: init: kumano-power: Cleanup and fix power parameters
This platform had "totally" bad configuration for power management:
first of all, we were trying to disable retention on L3, but...
surprise! the kernel does not even support it!

Then, there was a major issue about the bus-dcvs and entire
devfreq configuration for basically all of the internal devices:
we were * for some reason * trying to write by using wildcards,
which is not supported in Android init scripts: that was messing
up the entire platform sleep, locking the RPMh in active state.