vendor: mixer_paths: Add support for second record compress
The audio-record-compress2 path (second compress path for
concurrent compressed audio recording on the DSP) is supported
on the SM8150 SoC and by the HAL that we use right now... and
it's even getting requested for usage - RIGHT NOW! - so let's
enable this by implementing support for it.

Useless to say that - since it's being already requested - this
commit is not just a feature enhancement, but a solution for
various voice recording (outside of calls) issues.